--- src/app/components/pages/errorPage.tsx
@@ -1,14 +1,0 @@
-import React, { Component, ReactNode } from 'react';
-
-class ErrorPage extends Component {
-  render(): ReactNode {
-    return (
-      <>
-        <h1>404</h1>
-        <p>Page not found</p>
-      </>
-    );
-  }
-}
-
-export default ErrorPage;
