+++ src/app/components/pages/errorPage/errorPage.test.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { render, screen } from '@testing-library/react';
 import '@testing-library/jest-dom';
-import ErrorPage from '../app/components/pages/errorPage';
+import ErrorPage from './errorPage';
 
 describe('NavBar', () => {
   it('renders the navigation links and the current page', () => {
