+++ src/app/components/pages/errorPage/errorPage.tsx
@@ -1,0 +1,12 @@
+import React from 'react';
+
+const ErrorPage = () => {
+  return (
+    <>
+      <h1>404</h1>
+      <p>Page not found</p>
+    </>
+  );
+};
+
+export default ErrorPage;
